displaying two tables with incomes and expenses

+++ database/migrations/2024_03_09_115631_create_incomes_table.php
@@ -14,11 +14,15 @@ return new class extends Migration
         Schema::create('incomes', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('income_category_assigned_to_user_id');
+            $table->unsignedBigInteger('income_category_assigned_to_user_id');
             $table->decimal('amount', 8, 2, true);
             $table->date('date_of_income');
             $table->string('income_comment', 100);
             $table->timestamps();
+
+//            $table->foreign('income_category_assigned_to_user_id')->references('id')
+//                ->on('incomes_category_assigned_to_users');
+//            $table->foreignId('income_category_assigned_to_user_id')->constrained();
         });
     }
